tui/requirements.txt
textual
python-ulid
kombu
